% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_diff.R
\name{range_diff}
\alias{range_diff}
\title{Calculate differece}
\usage{
range_diff(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Range within the vector
}
\description{
Returns the range from the lowest to the highest value within a numeric vector
}
\examples{
range_diff(c(1, -1, 50, 99))
}
