% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate standard error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Standard error of the mean
}
\description{
This function calculates the standard error of the mean for a numeric vector
}
\examples{
standard_error(c(1, 3, 5, 7, 9))
}
